Move make_contract_class and make_compiled_class to contacts module
